refactor(session): extract shared types to session-types.ts and UI parsers to session-ui-parser.ts (#4228)
Recreated from current develop HEAD to avoid conflict resolution drift.

Changes:
- Extract UIState, SessionInfo, SessionState, PersistedStateData → session-types.ts
- Extract detectUIState, hasBlankPromptNearBottom, detectApprovalMethod → session-ui-parser.ts
- session.ts re-exports all types for backward compatibility
- Breaks circular deps: session-discovery, session-transcripts, monitor can import from session-types

No behavior changes — pure structural refactoring.
